# Switch to PrecompileTools.jl.