Testing `examples/kernel.jl` without a precompile workload at all:

```
❯ hyperfine -w 1 " julia --project examples/kernel.jl"
Benchmark 1:  julia --project examples/kernel.jl
  Time (mean ± σ):      8.131 s ±  0.241 s    [User: 8.139 s, System: 1.156 s]
  Range (min … max):    7.873 s …  8.375 s    10 runs
```

Current master:

```
Benchmark 1:  julia --project examples/kernel.jl
  Time (mean ± σ):      7.732 s ±  0.085 s    [User: 7.723 s, System: 1.173 s]
  Range (min … max):    7.560 s …  7.818 s    10 runs
```

Before https://github.com/JuliaGPU/GPUCompiler.jl/pull/600:

```
Benchmark 1:  julia --project examples/kernel.jl
  Time (mean ± σ):      1.782 s ±  0.022 s    [User: 1.825 s, System: 1.130 s]
  Range (min … max):    1.744 s …  1.815 s    10 runs
```

This PR:

```
Benchmark 1:  julia --project examples/kernel.jl
  Time (mean ± σ):     991.3 ms ±  14.8 ms    [User: 1045.8 ms, System: 1120.8 ms]
  Range (min … max):   968.7 ms … 1013.1 ms    10 runs
```

So still an improvement. Probably doesn't matter too much for downstream users, as the whole `@nospecialize(job::CompilerJob)` doesn't really work and I think we end up respecializing the GPUCompiler code anyway (i.e., downstream packages require their own precompile workload).