@vchuravy 2024-07-17
Yes but no... This is a tricky design issue/interaction.

I believe this is the same issue as why I want to use the custom method table for all GPU intrinsics and it should be equivalent to adding `precompile()`

Essentially precompilation tracking discards the CI and then uses the MI alone.