@vchuravy 2024-07-17
So it would be good to confirm if this is the same as:

https://github.com/JuliaLang/julia/pull/54155