@vchuravy 2024-07-17
Okay... Then we need to decide if this is the right fix... 
I wasn't convinced at the time.

Also just reminded me of https://github.com/JuliaLang/julia/pull/55072#discussion_r1669092995