@vchuravy 2024-07-17
> It's also surprising that this only triggers on Windows, no?

A little bit...